+++ lib/crdt.dart
@@ -3,5 +3,5 @@ library crdt;
 export 'src/crdt.dart';
 export 'src/crdt_json.dart';
 export 'src/hlc.dart';
-export 'src/map_crdt.dart';
+export 'src/crdt_map.dart';
 export 'src/record.dart';